requirements.txt
streamlit==1.35.0
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.4.2
plotly==5.22.0
faiss-cpu==1.7.4
sentence-transformers==2.7.0
google-generativeai==0.8.3
python-dotenv==1.0.0